Server/requirements.txt
fastapi==0.111.0
python-socketio[asyncio]>=5.11.0
uvicorn[standard]==0.29.0
redis==5.0.3
requests==2.31.0
python-multipart==0.0.9
